\name{cleanSingleLap}
\alias{cleanSingleLap}

\title{
Clean and generate dataframe for 1 lap
}

\description{
Returns a dataframe containing data collected from 1 lap
}
\usage{
cleanSingleLap(laponedata.csv, 1)
cleanSingleLap(laptwodata.csv, 2)
}

\arguments{
  \item{file}{a path to a csv file generated by racestudio2}
  \item{lapNum}{a number indicating the lap number of the data you are uploading, default value is 1}
}

\details{
}
\value{
A dataframe with data from racestudio2 that has been cleaned. A column for lap number has also been included based on the argument passed to lapNum.
}
\references{

}
\author{
Jess Kaminsky, Blain Morin, Fuyu Zou
}

\note{
}

\seealso{
}

\examples{
## The function is currently defined as
cleanSingleLap <- function(file, lapNum = 1) {
  lapData <- read_csv(file, skip = 15)

  lapData <- lapData[,-ncol(lapData)]

  nameData <-  read_csv(file, skip = 12)
  nameData <- nameData[,-ncol(nameData)]

  names(lapData) <- nameData[1,]
  lapData$Lap <- lapNum
  return(lapData)
  }
}

\keyword{clean}
\keyword{lap}
\keyword{data}
